redone the weird file hierarchy I had going
