+++ .gitignore
@@ -2,4 +2,3 @@
 .idea/
 
 classes/Keys.php
-/html/classes/Keys.php
