--- html/.gitignore
@@ -1,4 +1,0 @@
-.idea/*
-.idea/
-
-classes/Keys.php
